Temporary Memory Leak Fix
Payments contains an amounts array.  For whatever reason you generate about 300kb per payment (your UI displays 100 payments) equalling around 30mb.  You then save this to a giant in memory list.  This is bad.

I've lowered the amount of payments recorded to 3.  

There is a similar issue with workers but that is a fraction of this issue, I'll take a look at it if its a problem.